Move diagnostics/qc/qlp under kpfpipe.quality_control
Consolidate the three read-only post-processing layers into a single
kpfpipe.quality_control package and rename two of them:

  kpfpipe.qc  -> kpfpipe.quality_control.qc_binaries
  kpfpipe.qlp -> kpfpipe.quality_control.quicklook
  kpfpipe.diagnostics -> kpfpipe.quality_control.diagnostics

Updates all import paths across package internals, tests, scripts
(qc.py, qlp.py), and recipes (kpf_drp_science.py). No logic changes.
