--- kpfpipe/diagnostics/__init__.py
@@ -1,6 +1,0 @@
-from kpfpipe.diagnostics.base import Diagnostics
-from kpfpipe.diagnostics.level_0 import DiagL0
-from kpfpipe.diagnostics.level_1 import DiagL1
-from kpfpipe.diagnostics.level_2 import DiagL2
-
-__all__ = ["Diagnostics", "DiagL0", "DiagL1", "DiagL2"]
