--- kpfpipe/qc/__init__.py
@@ -1,6 +1,0 @@
-from kpfpipe.qc.base import QC
-from kpfpipe.qc.checks_l0 import QCL0
-from kpfpipe.qc.checks_l1 import QCL1
-from kpfpipe.qc.checks_l2 import QCL2
-
-__all__ = ["QC", "QCL0", "QCL1", "QCL2"]
