--- kpfpipe/qlp/__init__.py
@@ -1,4 +1,0 @@
-from kpfpipe.qlp.plot_l0 import PlotL0
-from kpfpipe.qlp.plot_l1 import PlotL1
-
-__all__ = ['PlotL0', 'PlotL1']
