+++ kpfpipe/quality_control/diagnostics/base.py
