+++ kpfpipe/quality_control/qc_binaries/checks_l0.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 from kpfpipe import REPO_ROOT
-from kpfpipe.qc.base import QC
+from kpfpipe.quality_control.qc_binaries.base import QC
 
 _JUNK_CSV = REPO_ROOT / "reference" / "junk_observations.csv"
