+++ kpfpipe/quality_control/qc_binaries/checks_l2.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from kpfpipe.qc.base import QC
+from kpfpipe.quality_control.qc_binaries.base import QC
 
 _CHIPS  = ["GREEN", "RED"]
 _FIBERS = ["SKY", "SCI1", "SCI2", "SCI3", "CAL"]
