+++ kpfpipe/quality_control/quicklook/__init__.py
@@ -1,0 +1,4 @@
+from kpfpipe.quality_control.quicklook.plot_l0 import PlotL0
+from kpfpipe.quality_control.quicklook.plot_l1 import PlotL1
+
+__all__ = ['PlotL0', 'PlotL1']
