+++ scripts/qc.py
@@ -19,7 +19,7 @@ import sys
 from kpfpipe.data_models.level0 import KPF0
 from kpfpipe.data_models.level1 import KPF1
 from kpfpipe.data_models.level2 import KPF2
-from kpfpipe.qc import QCL0, QCL1, QCL2
+from kpfpipe.quality_control.qc_binaries import QCL0, QCL1, QCL2
 from kpfpipe.utils.config import ConfigHandler
 from kpfpipe.utils.pipeline import build_filepath
